updatedCO and Smoke
